# api/generated.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

const (
	ApiKeyScopes     = "apiKey.Scopes"
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Actor defines model for Actor.
type Actor struct {
	OrgMsp string `json:"orgMsp"`
	UserId string `json:"userId"`
}

// CreateBusinessDataRequest defines model for CreateBusinessDataRequest.
type CreateBusinessDataRequest struct {
	// BusinessData Arbitrary JSON data
	BusinessData map[string]interface{} `json:"businessData"`
	RecordId     string                 `json:"recordId"`
}

// CreateInvoiceRequest defines model for CreateInvoiceRequest.
type CreateInvoiceRequest struct {
	Filename string `json:"filename"`
	RecordId string `json:"recordId"`

	// XmlBase64 Base64 encoded XML content
	XmlBase64 string `json:"xmlBase64"`
}

// Error defines model for Error.
type Error struct {
	Code    *int    `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// HistoryEntry defines model for HistoryEntry.
type HistoryEntry struct {
	IsDelete  *bool         `json:"isDelete,omitempty"`
	Timestamp *time.Time    `json:"timestamp,omitempty"`
	TxId      *string       `json:"txId,omitempty"`
	Value     *LedgerRecord `json:"value,omitempty"`
}

// LedgerRecord defines model for LedgerRecord.
type LedgerRecord struct {
	Actor Actor `json:"actor"`

	// BusinessData Can be generic JSON or InvoiceData structure
	BusinessData  *map[string]interface{} `json:"businessData,omitempty"`
	CreatedAt     *time.Time              `json:"createdAt,omitempty"`
	DocType       *string                 `json:"docType,omitempty"`
	LockPolicyId  *string                 `json:"lockPolicyId,omitempty"`
	Locked        *bool                   `json:"locked,omitempty"`
	LockedAt      *time.Time              `json:"lockedAt,omitempty"`
	PolicyVersion *int                    `json:"policyVersion,omitempty"`
	RecordId      string                  `json:"recordId"`
	Status        Status                  `json:"status"`
}

// Status defines model for Status.
type Status struct {
	Code      string     `json:"code"`
	Note      *string    `json:"note,omitempty"`
	UpdatedAt *time.Time `json:"updatedAt,omitempty"`
}

// UpdateBusinessDataRequest defines model for UpdateBusinessDataRequest.
type UpdateBusinessDataRequest struct {
	// BusinessData The new business data to replace the existing data
	BusinessData map[string]interface{} `json:"businessData"`
}

// UpdateInvoiceRequest defines model for UpdateInvoiceRequest.
type UpdateInvoiceRequest struct {
	NewFilename  string `json:"newFilename"`
	NewXmlBase64 string `json:"newXmlBase64"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Forbidden defines model for Forbidden.
type Forbidden = Error

// InternalServerError defines model for InternalServerError.
type InternalServerError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// CreateBusinessDataRecordJSONRequestBody defines body for CreateBusinessDataRecord for application/json ContentType.
type CreateBusinessDataRecordJSONRequestBody = CreateBusinessDataRequest

// CreateInvoiceRecordJSONRequestBody defines body for CreateInvoiceRecord for application/json ContentType.
type CreateInvoiceRecordJSONRequestBody = CreateInvoiceRequest

// UpdateBusinessDataJSONRequestBody defines body for UpdateBusinessData for application/json ContentType.
type UpdateBusinessDataJSONRequestBody = UpdateBusinessDataRequest

// UpdateInvoiceRecordJSONRequestBody defines body for UpdateInvoiceRecord for application/json ContentType.
type UpdateInvoiceRecordJSONRequestBody = UpdateInvoiceRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Health check
	// (GET /health)
	CheckHealth(w http.ResponseWriter, r *http.Request)
	// Create a record with generic business data
	// (POST /records/business-data)
	CreateBusinessDataRecord(w http.ResponseWriter, r *http.Request)
	// Create an Invoice record
	// (POST /records/invoice)
	CreateInvoiceRecord(w http.ResponseWriter, r *http.Request)
	// Get a ledger record by ID
	// (GET /records/{recordId})
	GetRecord(w http.ResponseWriter, r *http.Request, recordId string)
	// Update business data of a generic record
	// (PUT /records/{recordId}/business-data)
	UpdateBusinessData(w http.ResponseWriter, r *http.Request, recordId string)
	// Get history of a record
	// (GET /records/{recordId}/history)
	GetRecordHistory(w http.ResponseWriter, r *http.Request, recordId string)
	// Update an existing Invoice record
	// (PUT /records/{recordId}/invoice)
	UpdateInvoiceRecord(w http.ResponseWriter, r *http.Request, recordId string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Health check
// (GET /health)
func (_ Unimplemented) CheckHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a record with generic business data
// (POST /records/business-data)
func (_ Unimplemented) CreateBusinessDataRecord(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create an Invoice record
// (POST /records/invoice)
func (_ Unimplemented) CreateInvoiceRecord(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a ledger record by ID
// (GET /records/{recordId})
func (_ Unimplemented) GetRecord(w http.ResponseWriter, r *http.Request, recordId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update business data of a generic record
// (PUT /records/{recordId}/business-data)
func (_ Unimplemented) UpdateBusinessData(w http.ResponseWriter, r *http.Request, recordId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get history of a record
// (GET /records/{recordId}/history)
func (_ Unimplemented) GetRecordHistory(w http.ResponseWriter, r *http.Request, recordId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update an existing Invoice record
// (PUT /records/{recordId}/invoice)
func (_ Unimplemented) UpdateInvoiceRecord(w http.ResponseWriter, r *http.Request, recordId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// CheckHealth operation middleware
func (siw *ServerInterfaceWrapper) CheckHealth(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CheckHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateBusinessDataRecord operation middleware
func (siw *ServerInterfaceWrapper) CreateBusinessDataRecord(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateBusinessDataRecord(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateInvoiceRecord operation middleware
func (siw *ServerInterfaceWrapper) CreateInvoiceRecord(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateInvoiceRecord(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetRecord operation middleware
func (siw *ServerInterfaceWrapper) GetRecord(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "recordId" -------------
	var recordId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "recordId", runtime.ParamLocationPath, chi.URLParam(r, "recordId"), &recordId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "recordId", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRecord(w, r, recordId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// UpdateBusinessData operation middleware
func (siw *ServerInterfaceWrapper) UpdateBusinessData(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "recordId" -------------
	var recordId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "recordId", runtime.ParamLocationPath, chi.URLParam(r, "recordId"), &recordId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "recordId", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateBusinessData(w, r, recordId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetRecordHistory operation middleware
func (siw *ServerInterfaceWrapper) GetRecordHistory(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "recordId" -------------
	var recordId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "recordId", runtime.ParamLocationPath, chi.URLParam(r, "recordId"), &recordId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "recordId", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetRecordHistory(w, r, recordId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// UpdateInvoiceRecord operation middleware
func (siw *ServerInterfaceWrapper) UpdateInvoiceRecord(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "recordId" -------------
	var recordId string

	err = runtime.BindStyledParameterWithLocation("simple", false, "recordId", runtime.ParamLocationPath, chi.URLParam(r, "recordId"), &recordId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "recordId", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UpdateInvoiceRecord(w, r, recordId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.CheckHealth)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/records/business-data", wrapper.CreateBusinessDataRecord)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/records/invoice", wrapper.CreateInvoiceRecord)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/records/{recordId}", wrapper.GetRecord)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/records/{recordId}/business-data", wrapper.UpdateBusinessData)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/records/{recordId}/history", wrapper.GetRecordHistory)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/records/{recordId}/invoice", wrapper.UpdateInvoiceRecord)
	})

	return r
}